Migrate Preset classes to preset.py
Also forces imports to use load.dataset for FRDCDataset to avoid confusion and circular imports
