+++ Writerside/topics/preprocessing.extract_segments.md
@@ -135,7 +135,7 @@ Extract segments from bounds and labels.
 
 ```python
 import numpy as np
-from frdc.load import FRDCDataset
+from frdc.load.dataset import FRDCDataset
 from frdc.preprocess.extract_segments import extract_segments_from_bounds
 
 ds = FRDCDataset(site='chestnut_nature_park',
@@ -155,7 +155,7 @@ Extract segments from a label classification.
 from skimage.morphology import remove_small_objects, remove_small_holes
 import numpy as np
 
-from frdc.load import FRDCDataset
+from frdc.load.dataset import FRDCDataset
 from frdc.preprocess.morphology import (
     threshold_binary_mask, binary_watershed
 )
