--- src/frdc/load/__init__.py
@@ -1,3 +1,0 @@
-from .dataset import FRDCDataset
-
-__all__ = ["FRDCDataset"]
